// src/modules/meetings/server/types.ts
import { AppRouter } from "@/trpc/routers/_app";
import { inferRouterOutputs } from "@trpc/server";

export type MeetingGetOne = inferRouterOutputs<AppRouter>["meetings"]["getOne"];
export type MeetingGetMany = inferRouterOutputs<AppRouter>["meetings"]["getMany"]["items"]

export enum MeetingStatus {
  UPCOMING = "upcoming",
  COMPLETED = "completed",
  ACTIVE = "active",
  CANCELLED = "cancelled",
  PROCESSING = "processing"
}